Removed Cache package to because of circular package dependencies (which are bad and one way assess how good code is).
Fixed ProxyInvokers and CheckedInvokers so that they are Named Singletons, so that if two clients (java classes) ask for the CACHE_READER they actually get the same proxy, with out this updating the proxies would have had a issue where the last proxy asked for would be the only on in the methods map.  So if a programmer decided to rearrange a Registry during runtime, only part of it would actually get re arranged.